just trying something , everything work now

+++ 0x15-file_io/2-append_text_to_file.c
@@ -8,9 +8,9 @@
  */
 int append_text_to_file(const char *filename, char *text_content)
 {
-int file;
-int w;
-int str_len = 0;
+ssize_t file;
+ssize_t w;
+ssize_t str_len = 0;
 if (filename == NULL)
 {
 return (-1);
